Ask before removing

+++ djangocms_installer/main.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from distutils.util import strtobool
 import logging
 import os
 import shutil
@@ -34,19 +35,32 @@ def execute():
             if config_data.starting_page:
                 django.load_starting_page(config_data)
             if config_data.aldryn:
-                print("Project created!")
-                print("aldryn boilerplate requires action before you can actually run the project.\n"
-                      "See documentation at http://aldryn-boilerplate.readthedocs.org/ for more information.")
+                sys.stdout.write("Project created!\n")
+                sys.stdout.write("aldryn boilerplate requires action before you can actually run "
+                                 "the project.\n"
+                                 "See documentation at http://aldryn-boilerplate.readthedocs.org/"
+                                 "for more information.\n")
             else:
-                print("All done!")
-                print("Get into '%s' directory and type 'python manage.py runserver' "
-                      "to start your project" % config_data.project_directory)
+                sys.stdout.write("All done!\n")
+                sys.stdout.write("Get into '%s' directory and type 'python manage.py runserver' "
+                                 "to start your project\n" % os.path.abspath(config_data.project_directory))
     except Exception as e:
         # Clean up your own mess
         if os.path.exists(config_data.project_directory):
-            shutil.rmtree(config_data.project_directory, True)
+            choice = 'N'
+            if config_data.noinput is False:
+                sys.stdout.write("Failure occurred. Do you want to cleanup by removing %s? "
+                                 "[Y/N] " % os.path.abspath(config_data.project_directory))
+                choice = raw_input().lower()
+            else:
+                sys.stdout.write("Failure occurred.\n")
+            if strtobool(choice) or config_data.noinput:
+                sys.stdout.write("Removing everything under %s\n"
+                                 "" % os.path.abspath(config_data.project_directory))
+                shutil.rmtree(config_data.project_directory, True)
         if six.PY3:
             tb = sys.exc_info()[2]
-            raise EnvironmentError("%s\nDocumentation available at http://djangocms-installer.rtfd.org" % e).with_traceback(tb)
+            raise EnvironmentError("%s\nDocumentation available at "
+                                   "http://djangocms-installer.rtfd.org\n" % e).with_traceback(tb)
         else:
             raise
